address function comments
